pyproject.toml
[tool.poetry]
name = "petr-test-api"
version = "0.1.0"
description = ""
authors = ["Vladislav-76 <sillmaril2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.1.1"
allure-pytest = "^2.13.5"
requests = "^2.31.0"
allure-behave = "^2.13.5"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
